+++ src/types/cart.ts
@@ -1,0 +1,11 @@
+import { Product } from './product';
+
+export interface CartItem {
+  id: number;
+  quantity: number;
+  product: Product;
+}
+
+export interface NewCartItem extends CartItem {
+  quantity: 1;
+}
