+++ src/types/product.ts
@@ -4,13 +4,3 @@ export interface Product {
   price: number;
   imageUrl: string;
 }
-
-export interface CartItem {
-  id: number;
-  quantity: number;
-  product: Product;
-}
-
-export interface NewCartItem extends CartItem {
-  quantity: 1;
-}
